Add files via upload
Code seems alright and, wiring on the Arduino seems okay as well, but i don't know where to place a speaker on the Arduino.
The biggest problem is that the program, it doesn't run like i coded. After I applied this code to Arduino, water pump is just keep running forever. Can anyone check up my code to make this run properly ?